requirements.txt
python-dotenv
pymongo
sentence-transformers
ai21
PyPDF2
pandas
openpyxl
torch
scikit-learn
numpy
